--- src/main/java/me/taromati/streamerrecorder/FaviconConfig.java
@@ -1,32 +1,0 @@
-package me.taromati.streamerrecorder;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.Resource;
-import org.springframework.web.servlet.handler.SimpleUrlHandlerMapping;
-import org.springframework.web.servlet.resource.ResourceHttpRequestHandler;
-
-import java.util.Collections;
-import java.util.List;
-
-@Configuration
-public class FaviconConfig {
-
-    @Bean
-    public SimpleUrlHandlerMapping customFaviconHandlerMapping() {
-        SimpleUrlHandlerMapping mapping = new SimpleUrlHandlerMapping();
-        mapping.setOrder(Integer.MIN_VALUE);
-        mapping.setUrlMap(Collections.singletonMap(
-                "/favicon.ico", faviconRequestHandler()));
-        return mapping;
-    }
-
-    @Bean
-    protected ResourceHttpRequestHandler faviconRequestHandler() {
-        ResourceHttpRequestHandler requestHandler = new ResourceHttpRequestHandler();
-        List<Resource> locations = List.of(new ClassPathResource("/static/favicon/"));
-        requestHandler.setLocations(locations);
-        return requestHandler;
-    }
-}
