+++ src/main/java/me/taromati/streamerrecorder/discord/config/DiscordConfigProperties.java
@@ -1,0 +1,16 @@
+package me.taromati.streamerrecorder.discord.config;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Getter
+@Setter
+@Configuration
+@ConfigurationProperties(prefix = "discord")
+public class DiscordConfigProperties {
+    private String token;
+    private String serverName;
+    private String channelName;
+}
