+++ src/main/java/me/taromati/streamerrecorder/record/vo/SoopOption.java
@@ -1,0 +1,11 @@
+package me.taromati.streamerrecorder.record.vo;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class SoopOption {
+    private String username;
+    private String password;
+}
