--- src/main/java/me/taromati/streamerrecorder/vo/SoopOption.java
@@ -1,11 +1,0 @@
-package me.taromati.streamerrecorder.vo;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class SoopOption {
-    private String username;
-    private String password;
-}
